requirements.txt
-r tools/n64splat/requirements.txt
ansiwrap
anybadge
capstone
colorama
cxxfilt
pycparser
pylibyaml
pypng
python-Levenshtein
python-ranges
pyyaml
watchdog
spimdisasm==1.35.0
